move models and repository outside of api folder
