+++ backend/api/handlers/notificationHandler.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"github.com/Ygg-Drasill/PenaltyThing/backend/api/models"
+	"github.com/Ygg-Drasill/PenaltyThing/backend/models"
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
